FastAPI/requirements.txt
fastapi==0.110.0
python-multipart==0.0.9
uvicorn==0.29.0
annotated-types==0.6.0
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.124.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
pillow==10.2.0
python_dotenv==1.0.1
requests==2.31.0